* commit 'e972f13569e1870cc5e3d82bc4a58acb1eee6e27': (27 commits)
  update mdws configuration
  update sql scripts for CC, update var consul configuration
  add ECR-built mongo mocks and data seed, pull new images automatically.  Update README for additional relevant information.
  SM passwords configuration via Consul
  Adding HADB scripts
  README updates for running scripts and supplying parameters.
  add creation of var network to top of the stack
  breakdown of var nextgen local, allowing for a more segmented, modular approach to all components grouped by their roles in the overall stack, see updated README for more details
  consul configuration updates for var-resources
  use ECR mock mvi, leaving in option to build from local dev, update gitignore
  changed default to user mvi mock from iv
  Allow PROVIDER_NAME to be nullable
  MongoDB data changes to reflect a change in a data entry Added SQL Commit before beginning table alteration process in SQL script
  Upadting Telehealth service dependencies to AWS ECR
  Removed unnecessary certs Added and reused mongodb cert from PS Allowed connections without certificates for dev work
  Reorganized files for mongo volumes Added CA and certs Enabled SSL for mongo configurations
  VAR-10172 updated facility service and vmm to use mock mongo db
  Added cache cleanup of generated docker .cache Changed container host ports for mock-roa and var-oracle-db-mock Added existent name/tablespace exceptions to create tables script Changed default tablespace to respective VARDB_DATA/MOCKDB_DATA
  SCV-3071 Exposed var-oracle-db-mock 1521 port by unbinding mock-roa from host port 1521
  add mock for ALLINSTITUTIONS and sample data, rework drop script
  ...

+++ stop.sh
@@ -1,5 +1,32 @@
-echo -e "\nShutting down any NextGen running services…\n"
-docker-compose down || (echo "*** FAILED: Failed to stop running containers using docker-compose." && exit -1)
 
-echo -e "\nCleaning up dangling volumes"
-docker volume rm $(docker volume ls -f dangling=true -q)
+if [ -z $1 ] || [ "$1" == "all" ]; then
+    echo -e "\nShutting down NextGen platform services...\n"
+    docker-compose down || (echo "*** FAILED: Failed to stop running containers using docker-compose." && exit -1)
+
+    echo -e "\nCleaning up dangling volumes"
+    docker volume rm $(docker volume ls -f dangling=true -q)
+fi
+
+if [ "$1" == "fixtures" ] || [ "$1" == "all" ]; then
+    echo -e "\nShutting down fixture services...\n"
+    docker-compose -f docker-compose-fixtures.yml down
+fi
+
+if [ "$1" == "shared-services" ] || [ "$1" == "all" ]; then
+    echo -e "\nShutting down shared services...\n"
+    docker-compose -f docker-compose-shared-services.yml down
+fi
+
+if [ "$1" == "var" ] || [ "$1" == "all" ]; then
+    echo -e "\nShutting down var services and apps...\n"
+    docker-compose -f docker-compose-var.yml down
+fi
+
+if [ "$1" == "sm" ] || [ "$1" == "all" ]; then
+    echo -e "\nShutting down scheduling manager services and apps...\n"
+    docker-compose -f docker-compose-sm.yml down
+fi
+
+# removes cache file created by docker from var-oracle-db-mock
+echo -e "\nCleaning up generated docker cache file from oracle-db"
+rm ./var-oracle-db-mock/.cache
